Enable unbound configured to use an adblock list
The adblock list is imported from a github repository using a non-flake input to
this flake, and the program that compiles it into an unbound configuration is
provided by a flake.

Since that upstream flake is not regularly updated, we override its input of the
adblock list, pinning it in the flake.lock of the local repository instead.